refactor(typescript-sdk): split types/ + group internals/ to mirror Python SDK layout
The Python SDK has a clean per-domain layout (`types/task.py`,
`types/routing.py`, etc.) and groups implementation details under
their own modules. The TS SDK had drifted into a flat `src/` with one
237-line `types.ts` mixing six concerns (task / routing / verifier /
channels / task-handler / Phase 4 reserved) and implementation
helpers sitting alongside the public surface.

This commit fixes the layout. **No public API changes** — `index.ts`
re-exports everything in the same shape, so external consumers
(`import { awaitHuman } from "awaithumans"`) see no diff.

## Before

```
src/
  types.ts          # 237 lines, six domains
  await-human.ts
  constants.ts      # internal helper
  env.ts            # internal helper
  fetch.ts          # internal helper
  idempotency.ts    # internal helper
  wire.ts           # internal helper
  errors.ts
  index.ts
  reserved.ts
  schemas.ts
  testing.ts
  adapters/...
```

## After

```
src/
  index.ts          # public exports — unchanged shape
  await-human.ts
  errors.ts
  schemas.ts
  testing.ts
  reserved.ts
  types/                       # ← split per-domain
    index.ts                   # re-exports
    task.ts                    # AwaitHumanOptions, TaskRecord, TaskStatus
    routing.ts                 # AssignTo, HumanIdentity, Router (server-side ref)
    verification.ts            # VerifierConfig, VerifierResult, etc.
    channels.ts                # Channel, TaskTypeHandler (server-side ref)
  internal/                    # ← grouped implementation details
    constants.ts
    env.ts
    fetch.ts
    idempotency.ts
    wire.ts
  adapters/...
```

## What changed

  - `src/types.ts` (237 lines) → `src/types/{task,routing,verification,channels,index}.ts`
    (96/51/53/70/42 lines respectively)
  - Implementation helpers moved into `src/internal/`
  - Imports updated everywhere they referenced the old paths:
    `await-human.ts`, `errors.ts`, `adapters/temporal/index.ts`,
    `tests/idempotency.test.ts`
  - `internal/wire.ts` now imports `TaskStatus` via the package
    barrel rather than the previous flat `./types` reach-around

## What didn't change

  - Public API surface — every export from `index.ts` is still there
  - `dist/` shape mirrors the new src tree (verified via `tsc` build)
  - All 38 tests still pass

## Why this matters

Files over 200 lines are hard to hold in your head; the docs site
points at `types/task.py` for `AwaitHumanOptions` and the TS reader
should land in the same logical spot. New contributors have a
`internal/` signpost telling them what's helper vs. what's part of
the contract, instead of having to grep.
