+++ packages/typescript-sdk/src/internal/constants.ts
