+++ packages/typescript-sdk/src/internal/env.ts
